Add FactoryGirl declaration
