+++ lib/frontier/factory_girl_support.rb
@@ -2,3 +2,4 @@ class Frontier::FactoryGirlSupport
 end
 
 require_relative "./factory_girl_support/attributes_for"
+require_relative "./factory_girl_support/declaration"
