+++ lib/frontier/factory_girl_support/attributes_for.rb
@@ -7,17 +7,7 @@ class Frontier::FactoryGirlSupport::AttributesFor
   end
 
   def to_s
-    "FactoryGirl.attributes_for(:#{class_name})"
-  end
-
-private
-
-  def class_name
-    if model_configuration_or_association.is_a?(Frontier::ModelConfiguration)
-      model_configuration_or_association.model_name
-    elsif model_configuration_or_association.is_a?(Frontier::Association)
-      model_configuration_or_association.association_class.underscore
-    end
+    Frontier::FactoryGirlSupport::Declaration.new("attributes_for", model_configuration_or_association).to_s
   end
 
 end
